feat(mobile): centralize logging service with levels
