+++ mobile/app/_layout.tsx
@@ -18,7 +18,7 @@ import { useAutoLock } from '../hooks/useAutoLock';
 import { loadLanguage } from '../constants/i18n';
 import { getRouteFromNotificationData } from '../services/notifications/notificationRouting';
 import { biometricService } from '../services/security';
-import { logger } from '../utils/logger';
+import { logger } from '../services/logger';
 
 const ONBOARDING_KEY = 'onboardingComplete';
 const BIOMETRIC_LOCK_KEY = 'biometricLockEnabled';
